chore: accept "URI" as an alias for "LFN".
As we are introducing other storage types than lfn, namely girder as a
first step, the semantic meaning of the access type, in the
application configuration file for gasw, is really a uri.

The internal representation in gasw of this information is already a
uri, the enum type "GaswArgument.Type.URI".

The template in vip, gasw.vm, used to generate the gasw configuration
file can now be changed to use "URI" instead of "LFN".